Reorganize serialization-related modules into blaze.server.serialization
