+++ blaze/server/serialization/__init__.py
@@ -1,0 +1,7 @@
+from __future__ import absolute_import, division, print_function
+
+from .serialization import *
+from .json_dumps import json_dumps
+from .json_dumps_trusted import json_dumps_trusted
+from .object_hook import object_hook
+from .object_hook_trusted import object_hook_trusted
